requirements.txt
jenkinsapi>=0.3.4
pyserial>=3.3
